pyproject.toml
[project]
name = "friend-group"
version = "0.1.0"
description = "Virtual friend group chat powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]
